Remove web push commands
